fix: 경로 설정 수정 및 버셀 배포 에러 수정
