+++ src/components/Modals/Modal.module.css
@@ -107,10 +107,6 @@
 }
 
 /* Form 내부 간격 */
-/* .modal form > * {
-  margin-top: 32px;
-} */
-
 .cardForm > * {
   margin-top: 32px;
 }
