+++ src/components/Modals/domains/Card/Card.ts
