+++ src/components/Modals/domains/Card/DeleteCard.tsx
@@ -2,7 +2,7 @@
 
 import { useRouter } from 'next/navigation';
 import DefaultModal from '@/components/Modals/DefualtModal';
-import ModalButton from '@/components/Buttons/ModalButton/ModalButton';
+import ModalButton from '@/components/Buttons/shared/ModalButton/ModalButton';
 
 /**
  * DeleteCard 컴포넌트
