+++ src/components/Modals/domains/Invite/CreateInvite.tsx
@@ -3,8 +3,8 @@
 import { useRouter } from 'next/navigation';
 import DefaultModal from '@/components/Modals/DefualtModal';
 import { useState } from 'react';
-import ModalButton from '@/components/Buttons/ModalButton/ModalButton';
-import TextInput from '@/components/Input/TextInput/TextInput';
+import ModalButton from '@/components/Buttons/shared/ModalButton/ModalButton';
+import TextInput from '@/components/Input/domains/todo/TextInput/TextInput';
 
 type CreateInviteRequest = {
   email: string;
